requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.4
lxml==4.9.2
python-dotenv==1.0.0
requests==2.28.2
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.15
